[REFACTOR] 폴더링 변경
